init dev tools
