+++ .prettierrc.json
@@ -1,0 +1,7 @@
+{
+	"trailingComma": "none",
+	"useTabs": false,
+	"tabWidth": 2,
+	"semi": false,
+	"singleQuote": true
+}
