+++ .storybook/config.ts
@@ -1,0 +1,8 @@
+import { configure } from '@storybook/react'
+
+const req = require.context('../stories', true, /\.stories\.js$/)
+function loadStories() {
+  req.keys().forEach(filename => req(filename))
+}
+
+configure(loadStories, module)
