+++ .storybook/webpack.config.js
@@ -1,0 +1,18 @@
+module.exports = ({ config }) => {
+  config.module.rules.push({
+    test: /\.tsx?$/,
+    use: [
+      {
+        loader: require.resolve("babel-loader"),
+        options: {
+          presets: [require.resolve("babel-preset-react-app")]
+        }
+      },
+      require.resolve("react-docgen-typescript-loader")
+    ]
+  })
+
+  config.resolve.extensions.push(".ts", ".tsx")
+
+  return config
+}
