+++ package.json
@@ -6,17 +6,34 @@
   "repository": "https://github.com/kazushisan/shokken-ui.git",
   "author": "Kazushi Konosu <kazushi@konosu.me>",
   "license": "MIT",
-  "dependencies": {},
+  "dependencies": {
+    "react": "^16.8.6"
+  },
   "devDependencies": {
     "@babel/core": "^7.4.5",
     "@storybook/addon-actions": "^5.1.3",
+    "@storybook/addon-info": "^5.1.3",
     "@storybook/addon-links": "^5.1.3",
     "@storybook/addons": "^5.1.3",
     "@storybook/react": "^5.1.3",
-    "babel-loader": "^8.0.6"
+    "@types/storybook__react": "^4.0.2",
+    "@typescript-eslint/eslint-plugin": "^1.10.2",
+    "@typescript-eslint/parser": "^1.10.2",
+    "babel-loader": "^8.0.6",
+    "eslint": "^5.16.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-config-prettier": "^4.3.0",
+    "eslint-plugin-import": "^2.17.3",
+    "eslint-plugin-jsx-a11y": "^6.2.1",
+    "eslint-plugin-prettier": "^3.1.0",
+    "eslint-plugin-react": "^7.13.0",
+    "prettier": "^1.18.2",
+    "react-docgen-typescript-loader": "^3.1.0",
+    "typescript": "^3.5.1"
   },
   "scripts": {
     "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "build-storybook": "build-storybook",
+    "lint": "eslint --ext .ts,tsx ./"
   }
 }
