+++ tsconfig.json
@@ -1,0 +1,15 @@
+{
+	"compilerOptions": {
+		"strict": true,
+		"sourceMap": true,
+		"noImplicitReturns": true,
+		"noUnusedLocals": true,
+		"target": "es5",
+		"module": "es2015",
+		"jsx": "react",
+		"moduleResolution": "node",
+		"experimentalDecorators": true,
+		"lib": ["dom", "es6"],
+		"allowSyntheticDefaultImports": true
+	}
+}
